+++ pollbot/views/release.py
@@ -44,13 +44,17 @@ devedition_beta_check = status_response(devedition_and_beta_in_sync)
 balrog_rules = status_response(balrog.balrog_rules)
 buildhub_check = status_response(buildhub.buildhub)
 crash_stats_uptake = status_response(crash_stats.uptake)
-telemetry_uptake = status_response(telemetry.update_parquet_uptake)
+telemetry_restart_after_update = status_response(telemetry.restart_after_update)
+telemetry_migrated_from_previous_version = status_response(
+    telemetry.migrated_from_previous_version)
 
 
 @validate_product_version
 async def view_get_releases(request, product):
+    releases = await get_releases(product)
+    releases = sorted(set([r[1] for r in releases]), key=lambda version: build_version_id(version))
     return web.json_response({
-        "releases": await get_releases(product)
+        "releases": releases
     })
 
 
@@ -66,7 +70,8 @@ CHECKS_TITLE = {
     "balrog-rules": "Balrog update rules",
     "buildhub": "Buildhub release info",
     "crash-stats-uptake": "Crash Stats Uptake (24h latency)",
-    "telemetry-update-parquet-uptake": "Telemetry Update Parquet Uptake",
+    "telemetry-restart-after-update": "Telemetry - People having restarted after update download",
+    "telemetry-migrated": "Telemetry - People having migrated from previous version",
 }
 
 ALL = [Channel.ESR, Channel.RELEASE, Channel.CANDIDATE,
@@ -89,10 +94,11 @@ CHECKS = OrderedDict(
         "bouncer": [Channel.ESR, Channel.RELEASE, Channel.BETA, Channel.AURORA, Channel.NIGHTLY],
         "buildhub": ALL,
         "crash-stats-uptake": [Channel.ESR, Channel.RELEASE, Channel.BETA],
-        "telemetry-update-parquet-uptake": "57.0a1",
+        "telemetry-restart-after-update": "57.0a1",
+        "telemetry-migrated": "57.0a1",
     }.items(), key=lambda t: t[0]))
 
-NOT_ACTIONABLE = ['-uptake']
+NOT_ACTIONABLE = ['-uptake', 'telemetry-']
 IGNORES = {'devedition': ['crash-stats-uptake', 'partner-repacks']}