+++ CHANGELOG.rst
@@ -5,7 +5,9 @@ CHANGELOG
 ------------------
 
 - Order checks putting non actionable checks at the end. (#163)
-- Fix Telemetry query. Refs `Bug 1429048 <https://bugzilla.mozilla.org/show_bug.cgi?id=1429048#c9>`_
+- Fix Telemetry "percentage of users having restarted after automatic update".
+  Refs `Bug 1429048 <https://bugzilla.mozilla.org/show_bug.cgi?id=1429048#c9>`_
+- Add Telemetry "percentage of users having migrated"
 
 
 0.6.1 (2017-12-20)